add jsdoc, refactor
