+++ packages/client/package.json
@@ -30,6 +30,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^28",
     "jest-environment-jsdom": "^29.0.1",
+    "jsdoc": "^4.0.4",
     "lefthook": "^1.3.9",
     "prettier": "^2.7.1",
     "ts-jest": "^28.0.8",
