+++ packages/client/src/entities/error/error.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Button } from '../button'
+import { Button } from '../../shared/button'
 
 import styles from './styles.module.css'
