+++ packages/client/src/entities/error/index.ts
