+++ packages/client/src/entities/error/styles.module.css
