+++ packages/client/src/shared/button/button.tsx
@@ -4,12 +4,29 @@ import classNames from 'classnames'
 import styles from './styles.module.css'
 
 type TButtonProps = {
+  /**
+   * Размер кнопки
+   * @default 's'
+   */
   size?: 's' | 'm' | 'l' | 'xl'
-  color: 'primary' | 'secondary' | 'contrast'
-  label: string
+  /**
+   * Цвет кнопки
+   * @default 'primary'
+   */
+  color?: 'primary' | 'secondary' | 'contrast'
+  /**
+   * Текст кнопки
+   * @default 'label'
+   */
+  label?: string
 } & React.HTMLAttributes<HTMLButtonElement>
 
-export const Button = ({ size = 's', color, label, onClick }: TButtonProps) => {
+export const Button = ({
+  size = 's',
+  color = 'primary',
+  label = 'label',
+  onClick,
+}: TButtonProps) => {
   const className = classNames(styles.buttonContainer, {
     [styles.buttonS]: size === 's',
     [styles.buttonM]: size === 'm',
